- Fixed watcher - Added old animation options - Removed matrix - Added underscore strings

+++ scripts/watch.coffee
@@ -1,17 +1,21 @@
 watchr = require "watchr"
 {spawn} = require "child_process"
 
-path = process.argv[2]
-cmd = process.argv[3]
+paths = process.argv[2].split ","
+command = process.argv[3]
+commandArguments = process.argv[4..]
 
 child = null
 
 runCommand = ->
 	
-	console.log "Running '#{cmd} #{process.argv[4..].join " "}'"
+	if not command
+		return console.log "Missing command"
+
+	console.log "Running '#{command} #{process.argv[4..].join " "}'"
 
 	child?.kill()
-	child = spawn cmd, process.argv[4..]
+	child = spawn command, commandArguments
 
 	child.stdout.on "data",  (data) -> process.stdout.write Buffer(data).toString()
 	child.stdout.on "error", (data) -> process.stderr.write Buffer(data).toString()
@@ -19,19 +23,17 @@ runCommand = ->
 	child.stderr.on "data",  (data) -> process.stdout.write Buffer(data).toString()
 	child.stderr.on "error", (data) -> process.stderr.write Buffer(data).toString()
 
-console.log "Path: #{path}"
-console.log "Command: #{cmd}"
 
 runCommand()
 
 watchr.watch
-	paths: [path]
+	paths: paths
+	catchupDelay: 100
+	persistent: false
+	ignoreHiddenFiles: true
 	listeners:
-		error: (err) ->
-			console.log "an error occured:", err
-
+		error: (err) -> console.log "an error occured:", err
 		change: (changeType, filePath) ->
-			
 			if filePath.indexOf(".coffee") > -1
 				console.log "Change: #{filePath}"
 				runCommand()
